+++ data/data.json
@@ -487,76 +487,22 @@
   ],
   "users": [
     {
-      "id": "dcca",
-      "email": "q@a.com",
-      "password": "U29tZXRoaW5nOSE=",
-      "avatar": "../img/avatar.jpg"
-    },
-    {
-      "id": "2d34",
-      "email": "qq@a.com",
-      "password": "U29tZXRoaW5nOSE=",
-<<<<<<< HEAD
-      "avatar": "../img/avatar.jpg"
-=======
-      "avatar": "/img/avatar.jpg"
->>>>>>> 2aed1a0 (initial commit)
-    },
-    {
-      "id": "3b29",
-      "email": "q@a1.com",
-      "password": "U29tZXRoaW5nOSE=",
-<<<<<<< HEAD
-      "avatar": "../img/avatar.jpg"
-=======
-      "avatar": "/img/avatar.jpg"
->>>>>>> 2aed1a0 (initial commit)
-    },
-    {
+      "id": "f180",
       "email": "something@email.com",
-<<<<<<< HEAD
-      "password": "U29tZXRoaW5nOSE=",
-<<<<<<< HEAD
-      "avatar": "../img/avatar.jpg"
-=======
-      "avatar": "/avatar.jpg"
->>>>>>> 2aed1a0 (initial commit)
-=======
-      "avatar": "https://replicate.delivery/pbxt/JF3foGR90vm9BXSEXNaYkaeVKHYbJPinmpbMFvRtlDpH4MMk/out-0-1.png",
-      "password": "U29tZXRoaW5nMTEh",
-      "id": "88ca"
->>>>>>> 9e5b70c (update modal functional, need tweaking)
+      "password": "U29tZXRoaW5nMTAh",
+      "avatar": "https://tinypng.com/images/social/website.jpg"
     },
     {
-      "id": "36dc",
-      "email": "Something9@some.com",
+      "id": "3532",
+      "email": "something1@email.com",
       "password": "U29tZXRoaW5nOSE=",
-<<<<<<< HEAD
-      "avatar": "../img/avatar.jpg"
-=======
-      "avatar": "/img/avatar.jpg"
+      "avatar": "https://gratisography.com/wp-content/uploads/2024/10/gratisography-cool-cat-800x525.jpg"
     },
     {
-      "id": "0021",
-      "email": "something@email1.com",
-      "password": "U29tZXRoaW5nOSE=",
-      "avatar": "/avatar.jpg"
-    },
-    {
-      "id": "dcda",
+      "id": "823c",
       "email": "something2@email.com",
       "password": "U29tZXRoaW5nOSE=",
-      "avatar": "/img/avatar.jpg"
-<<<<<<< HEAD
->>>>>>> 2aed1a0 (initial commit)
-=======
-    },
-    {
-      "id": "39d4",
-      "email": "new@someone.com",
-      "password": "U29tZXRoaW5nOSE=",
       "avatar": "/avatar.jpg"
->>>>>>> 9e5b70c (update modal functional, need tweaking)
     }
   ]
 }
