+++ src/assets/components/messages/SignUpSuccess.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { useData } from "./DataContext";
+import { useData } from "../DataContext";
 
 const SignUpSuccess = () => {
   const { success } = useData();
@@ -24,7 +24,7 @@ const SignUpSuccess = () => {
         {success && (
           <div className="flex justify-center items-center">
             <div
-              className={`bg-figma-semi-dark-blue tablet:w-[25rem] tablet:h-[6rem] phone:w-[20.4375rem] phone:h-[6rem] flex justify-center items-center rounded-[1.25rem] text-figma-white figma-body-m border border-green-400 shadow-md shadow-green-400 fixed bottom-[10rem] transition-all duration-500  ${
+              className={`z-[100] bg-figma-semi-dark-blue tablet:w-[25rem] tablet:h-[6rem] phone:w-[20.4375rem] phone:h-[6rem] flex justify-center items-center rounded-[1.25rem] text-figma-white figma-body-m border border-green-400 shadow-md shadow-green-400 fixed bottom-[10rem] transition-all duration-500  ${
                 isVisible
                   ? "translate-y-0 opacity-100"
                   : "translate-y-full opacity-0"
